+++ src/api/runtime.ts
@@ -15,9 +15,6 @@ export function makeRuntime({
   pickleIds,
   newId,
   supportCodeLibrary,
-  requireModules,
-  requirePaths,
-  importPaths,
   options: { parallel, ...options },
 }: {
   cwd: string
@@ -27,9 +24,6 @@ export function makeRuntime({
   newId: IdGenerator.NewId
   pickleIds: string[]
   supportCodeLibrary: SupportCodeLibrary
-  requireModules: string[]
-  requirePaths: string[]
-  importPaths: string[]
   options: IRunOptionsRuntime
 }): IRuntime {
   if (parallel > 0) {
@@ -42,9 +36,6 @@ export function makeRuntime({
       options,
       newId,
       supportCodeLibrary,
-      requireModules,
-      requirePaths,
-      importPaths,
       numberOfWorkers: parallel,
     })
   }