+++ src/runtime/parallel/worker.ts
@@ -1,5 +1,6 @@
 import { EventEmitter } from 'node:events'
 import { pathToFileURL } from 'node:url'
+import { register } from 'node:module'
 import * as messages from '@cucumber/messages'
 import { IdGenerator } from '@cucumber/messages'
 import { JsonObject } from 'type-fest'
@@ -53,36 +54,32 @@ export default class Worker {
     this.sendMessage = sendMessage
     this.eventBroadcaster = new EventEmitter()
     this.eventBroadcaster.on('envelope', (envelope: messages.Envelope) => {
-      // assign `workerId` property only for the `testCaseStarted` message
-      if (envelope.testCaseStarted) {
-        envelope.testCaseStarted.workerId = this.id
-      }
-      this.sendMessage({ jsonEnvelope: JSON.stringify(envelope) })
+      this.sendMessage({ jsonEnvelope: envelope })
     })
   }
 
   async initialize({
-    filterStacktraces,
-    requireModules,
-    requirePaths,
-    importPaths,
+    supportCodeCoordinates,
     supportCodeIds,
     options,
   }: IWorkerCommandInitialize): Promise<void> {
-    supportCodeLibraryBuilder.reset(this.cwd, this.newId, {
-      requireModules,
-      requirePaths,
-      importPaths,
-    })
-    requireModules.map((module) => tryRequire(module))
-    requirePaths.map((module) => tryRequire(module))
-    for (const path of importPaths) {
+    supportCodeLibraryBuilder.reset(
+      this.cwd,
+      this.newId,
+      supportCodeCoordinates
+    )
+    supportCodeCoordinates.requireModules.map((module) => tryRequire(module))
+    supportCodeCoordinates.requirePaths.map((module) => tryRequire(module))
+    for (const specifier of supportCodeCoordinates.loaders) {
+      register(specifier, pathToFileURL('./'))
+    }
+    for (const path of supportCodeCoordinates.importPaths) {
       await import(pathToFileURL(path).toString())
     }
     this.supportCodeLibrary = supportCodeLibraryBuilder.finalize(supportCodeIds)
 
     this.worldParameters = options.worldParameters
-    this.filterStacktraces = filterStacktraces
+    this.filterStacktraces = options.filterStacktraces
     this.runTestRunHooks = makeRunTestRunHooks(
       options.dryRun,
       this.supportCodeLibrary.defaultTimeout,
@@ -125,6 +122,7 @@ export default class Worker {
   }: IWorkerCommandRun): Promise<void> {
     const stopwatch = create(elapsed)
     const testCaseRunner = new TestCaseRunner({
+      workerId: this.id,
       eventBroadcaster: this.eventBroadcaster,
       stopwatch,
       gherkinDocument,